fix(sdk): one formatUsdc instead of two independent copies
parseUsdcAmount was already shared from tools/parse/usdcAmount.ts. Its
inverse, formatUsdc, was copied into buildCctpBridge.ts and
threeJane/buildSupplyUsdc.ts independently.

The two copies still agreed, so nothing is broken today. What makes it worth
fixing rather than noting: formatUsdc produces the `amountUsdc` string a user
READS on a signing card. A drift between the copies renders the same amount
two different ways depending on which builder produced it - and a display-only
divergence on a fund path is the kind that gets believed.

Moved it next to its inverse and pointed both consumers at it. Exported from
the threeJane barrel too, matching how parseUsdcAmount is already surfaced
there; the tools/bridge export path is unchanged, so no consumer import moves.

The test mirrors the existing shared-parseUsdcAmount one: identity, not
behaviour. Asserting both render '1.5' would pass with two copies still in
place - asserting they are the SAME FUNCTION is what makes a future local copy
fail.
