+++ packages/sdk/src/tools/parse/usdcAmount.ts
@@ -32,3 +32,23 @@ export const parseUsdcAmount = (value: string): bigint => {
   const fractionalUnits = fractionalPart === '' ? 0n : BigInt(fractionalPart)
   return BigInt(wholePart) * 10n ** BigInt(USDC_DECIMALS) + fractionalUnits
 }
+
+/**
+ * Render raw 6-decimal USDC units as a human-readable amount. The exact inverse
+ * of {@link parseUsdcAmount}: trailing fractional zeros are dropped, and a whole
+ * amount renders without a decimal point at all.
+ *
+ * Lives here, next to its inverse, because it was previously copied into
+ * `buildCctpBridge.ts` and `threeJane/buildSupplyUsdc.ts` independently
+ * (sdk#1931). The two copies happened to still agree, but they are what
+ * produces the `amountUsdc` string a user reads on a signing card, so a drift
+ * between them would show the same amount two different ways depending on which
+ * builder produced it.
+ */
+export const formatUsdc = (raw: bigint): string => {
+  const divisor = 10n ** BigInt(USDC_DECIMALS)
+  const whole = raw / divisor
+  const frac = raw % divisor
+  if (frac === 0n) return whole.toString()
+  return `${whole}.${frac.toString().padStart(USDC_DECIMALS, '0').replace(/0+$/, '')}`
+}
